projects/capstone/backend/requirements.txt
autopep8==2.0.0
click==8.1.3
ecdsa==0.18.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.9.1
python-dotenv==0.21.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
